+++ git-repository/src/ext/mod.rs
@@ -1,0 +1,8 @@
+mod tree;
+pub use tree::TreeIterExt;
+
+mod object_id;
+pub use object_id::ObjectIdExt;
+
+pub(crate) mod access;
+pub use access::{object::ObjectAccessExt, reference::ReferenceAccessExt};
