+++ git-repository/src/ext/object_id.rs
@@ -1,0 +1,34 @@
+use git_hash::ObjectId;
+#[cfg(feature = "git-traverse")]
+use git_object::immutable;
+#[cfg(feature = "git-traverse")]
+use git_traverse::commit::ancestors::{Ancestors, State};
+
+use crate::{Access, Oid};
+
+pub trait Sealed {}
+
+pub trait ObjectIdExt: Sealed {
+    #[cfg(feature = "git-traverse")]
+    fn ancestors_iter<Find>(self, find: Find) -> Ancestors<Find, fn(&git_hash::oid) -> bool, State>
+    where
+        Find: for<'a> FnMut(&git_hash::oid, &'a mut Vec<u8>) -> Option<immutable::CommitIter<'a>>;
+
+    fn attach<A: Access + Sized>(self, access: &A) -> Oid<'_, A>;
+}
+
+impl Sealed for ObjectId {}
+
+impl ObjectIdExt for ObjectId {
+    #[cfg(feature = "git-traverse")]
+    fn ancestors_iter<Find>(self, find: Find) -> Ancestors<Find, fn(&git_hash::oid) -> bool, State>
+    where
+        Find: for<'a> FnMut(&git_hash::oid, &'a mut Vec<u8>) -> Option<immutable::CommitIter<'a>>,
+    {
+        Ancestors::new(Some(self), State::default(), find)
+    }
+
+    fn attach<A: Access + Sized>(self, access: &A) -> Oid<'_, A> {
+        Oid::from_id(self, access)
+    }
+}
